perf(db): otimizar chave primária de dados_farmacia para INT
- Altera o tipo da coluna 'id' em fp.dados_farmacia de BIGINT para INT.
- Otimização visa reduzir o overhead de memória e disco em ~4GB no processamento
  da tabela mestre de movimentação (1 bilhão+ de linhas).
- Prepara a infraestrutura para a migração da arquitetura de identificação
  de farmácias de CNPJ (String) para Surrogate Key numérico.

+++ src/scripts/preparação_dados.sql
@@ -377,7 +377,7 @@ CROSS APPLY (
 ) AS outras;
 
 -- Colunas de controle e georreferenciamento
-ALTER TABLE #tempDadosFarmacias2 ADD id        BIGINT IDENTITY;
+ALTER TABLE #tempDadosFarmacias2 ADD id        INT IDENTITY;
 ALTER TABLE #tempDadosFarmacias2 ADD latitude  DECIMAL(9, 6);
 ALTER TABLE #tempDadosFarmacias2 ADD longitude DECIMAL(9, 6);
